Rename styled folder
