+++ src/components/Forms/SignInForm/SignInForm.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { StyledFormLayout, StyledFormPattern, StyledFormWrap } from '../Form/style';
+import { StyledFormLayout, StyledFormPattern, StyledFormWrap } from '../Form/styled';
 
 const SignInForm = (props) => {
     const { children } = props;
