+++ src/pages/Home/styled.ts
